Add to registry fix
